[XCConfig] Collapse public and private pod xcconfig
