+++ lib/cocoapods/generator/xcconfig.rb
@@ -6,8 +6,7 @@ module Pod
     #
     module XCConfig
       autoload :AggregateXCConfig,  'cocoapods/generator/xcconfig/aggregate_xcconfig'
-      autoload :PublicPodXCConfig,  'cocoapods/generator/xcconfig/public_pod_xcconfig'
-      autoload :PrivatePodXCConfig, 'cocoapods/generator/xcconfig/private_pod_xcconfig'
+      autoload :PodXCConfig,        'cocoapods/generator/xcconfig/pod_xcconfig'
       autoload :XCConfigHelper,     'cocoapods/generator/xcconfig/xcconfig_helper'
     end
   end
