+++ lib/cocoapods/generator/xcconfig/pod_xcconfig.rb
@@ -3,28 +3,21 @@ module Pod
     module XCConfig
       # Generates the private xcconfigs for the pod targets.
       #
-      # The private xcconfig file for a Pod target merges the configuration
-      # values of the public namespaced xcconfig with the default private
-      # configuration values required by CocoaPods.
+      # The xcconfig file for a Pod target merges the pod target
+      # configuration values with the default configuration values
+      # required by CocoaPods.
       #
-      class PrivatePodXCConfig
+      class PodXCConfig
         # @return [Target] the target represented by this xcconfig.
         #
         attr_reader :target
 
-        # @return [Xcodeproj::Config] The public xcconfig which this one will
-        #         use.
-        #
-        attr_reader :public_xcconfig
-
         # Initialize a new instance
         #
         # @param  [Target] target @see target
-        # @param  [Xcodeproj::Config] public_xcconfig @see public_xcconfig
         #
-        def initialize(target, public_xcconfig)
+        def initialize(target)
           @target = target
-          @public_xcconfig = public_xcconfig
         end
 
         # @return [Xcodeproj::Config] The generated xcconfig.
@@ -71,58 +64,10 @@ module Pod
             config.merge!(build_settings)
           end
 
-          xcconfig_hash = add_xcconfig_namespaced_keys(public_xcconfig.to_hash, config, target.xcconfig_prefix)
-          @xcconfig = Xcodeproj::Config.new(xcconfig_hash)
+          @xcconfig = Xcodeproj::Config.new(config)
+          XCConfigHelper.add_settings_for_file_accessors_of_target(target, @xcconfig)
           XCConfigHelper.add_target_specific_settings(target, @xcconfig)
-          @xcconfig.includes = [target.name]
           @xcconfig
-        end
-
-        private
-
-        #-----------------------------------------------------------------------#
-
-        # !@group Private Helpers
-
-        # Returns the hash representation of an xcconfig which inherit from the
-        # namespaced keys of a given one.
-        #
-        # @param  [Hash] source_config
-        #         The xcconfig whose keys need to be inherited.
-        #
-        # @param  [Hash] destination_config
-        #         The config which should inherit the source config keys.
-        #
-        # @return [Hash] The inheriting xcconfig.
-        #
-        def add_xcconfig_namespaced_keys(source_config, destination_config, prefix)
-          result = destination_config.dup
-          source_config.each do |key, _value|
-            prefixed_key = prefix + conditional_less_key(key)
-            current_value = destination_config[key]
-            if current_value
-              result[key] = "#{current_value} ${#{prefixed_key}}"
-            else
-              result[key] = "${#{prefixed_key}}"
-            end
-          end
-          result
-        end
-
-        # Strips the [*]-syntax from the given xcconfig key.
-        #
-        # @param  [String] key
-        #         The key to strip.
-        #
-        # @return [String] The stripped key.
-        #
-        def conditional_less_key(key)
-          brackets_index = key.index('[')
-          if brackets_index
-            key[0...brackets_index]
-          else
-            key
-          end
         end
 
         #-----------------------------------------------------------------------#
