+++ spec/unit/installer/target_installer/pod_target_installer_spec.rb
@@ -100,7 +100,6 @@ module Pod
           @installer.install!
           group = @project['Pods/BananaLib/Support Files']
           group.children.map(&:display_name).sort.should == [
-            'Pods-BananaLib-Private.xcconfig',
             'Pods-BananaLib-dummy.m',
             'Pods-BananaLib-prefix.pch',
             'Pods-BananaLib.xcconfig',
@@ -127,7 +126,7 @@ module Pod
           @installer.install!
           bundle_target = @project.targets.find { |t| t.name == 'Pods-BananaLib-banana_bundle' }
 
-          file = config.sandbox.root + @pod_target.xcconfig_private_path
+          file = config.sandbox.root + @pod_target.xcconfig_path
           bundle_target.build_configurations.each do |bc|
             bc.base_configuration_reference.real_path.should == file
           end
@@ -142,7 +141,6 @@ module Pod
           @project.support_files_group
           group = @project['Pods/BananaLib/Support Files']
           group.children.map(&:display_name).sort.should == [
-            'BananaLib-Private.xcconfig',
             'BananaLib-dummy.m',
             'BananaLib-prefix.pch',
             'BananaLib.xcconfig',
@@ -169,7 +167,7 @@ module Pod
           @installer.install!
           bundle_target = @project.targets.find { |t| t.name == 'BananaLib-banana_bundle' }
 
-          file = config.sandbox.root + @pod_target.xcconfig_private_path
+          file = config.sandbox.root + @pod_target.xcconfig_path
           bundle_target.build_configurations.each do |bc|
             bc.base_configuration_reference.real_path.should == file
           end
@@ -180,7 +178,7 @@ module Pod
 
       it 'creates the xcconfig file' do
         @installer.install!
-        file = config.sandbox.root + @pod_target.xcconfig_private_path
+        file = config.sandbox.root + @pod_target.xcconfig_path
         xcconfig = Xcodeproj::Config.new(file)
         xcconfig.to_hash['PODS_ROOT'].should == '${SRCROOT}'
       end
